module/requirements.txt
requests
PyMySQL==0.9.3
python-dotenv==0.11.0
Flask==1.1.1
opencc-python-reimplemented==0.1.5
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
APScheduler==3.5.0
pymessenger==0.0.7.0